make BayerLuetticke importable and rewrite TwoAssets
