+++ HARK/BayerLuetticke/OneAssetCode_HANK/EX2SS.p
