+++ HARK/BayerLuetticke/OneAssetCode_HANK/SharedFunc2.py
