+++ HARK/BayerLuetticke/OneAssetCode_HANK/defineSSParameters.py
