+++ HARK/BayerLuetticke/OneAssetCode_KS/EX1SS.p
