+++ HARK/BayerLuetticke/OneAssetCode_KS/EX1SS_nm50.p
