+++ HARK/BayerLuetticke/OneAssetCode_KS/SharedFunc.py
