+++ HARK/BayerLuetticke/TwoAssetCode/EX3SS_20.p
