+++ HARK/BayerLuetticke/TwoAssetCode/SharedFunc3.py
