+++ HARK/BayerLuetticke/TwoAssetCode/defineSSParametersTwoAsset.py
