+++ HARK/BayerLuetticke/notebooks/DCT-Copula-Illustration.py
