+++ HARK/BayerLuetticke/notebooks/OneAsset-HANK-MP.py
