+++ HARK/__init__.py
@@ -2,3 +2,5 @@ from __future__ import absolute_import
 from .core import *
 
 __version__ = '0.10.1'
+
+from .BayerLuetticke import *
